// packages/component/src/block/tradingView/types.ts
export type TimeInterval =
  | "1"
  | "3"
  | "5"
  | "15"
  | "30"
  | "60"
  | "120"
  | "240"
  | "720"
  | "1D"
  | "1W";